add Example005 If-Else and fix Example004

+++ Example004_Div/Program.cs
@@ -1,5 +1,5 @@
 ﻿// Деление
 
-double numberA = 12;
-double numberB = 5;
+int numberA = 12;
+int numberB = 5;
 Console.WriteLine(numberA / numberB);
